Move time spent summary function to utils
