+++ app/src/popup/Utils.ts
@@ -1,3 +1,5 @@
+import Database from "../engine/Database";
+
 export interface Icon {
     size: number;
     src: string;
@@ -23,6 +25,21 @@ export const getWebsiteIconObject = (websiteURL: string | undefined): Icon => {
     };
     return icon;
 };
+
+export function storeTimeSpentSummary(currentDomain: string) {
+    const db = new Database();
+    const previousDomain = db.readPreviousDomain();
+
+    if (currentDomain.length > 0 && previousDomain !== currentDomain) {
+        db.writePreviousDomain(currentDomain);
+        db.writeLastActive(currentDomain, new Date());
+
+        const lastActive = db.readLastActive(previousDomain);
+        const timeSpent = Math.trunc(Math.abs(Date.now() - lastActive.getTime()) / 1000);
+        const totalTimeSpentOfLastActive = (db.readTimeSpent(previousDomain) as number) + timeSpent;
+        db.writeTimeSpent(previousDomain, totalTimeSpentOfLastActive);
+    }
+}
 
 export const howManyHoursInSeconds = (timeInSeconds: number): number => {
     const secondsInOneHour = 60 * 60;
